refactor(image): Move sharp's parseParams to baseService
