+++ packages/astro/src/image/services/sharp.ts
@@ -1,4 +1,4 @@
-import { ImageTransform, OutputFormat } from '../types.js';
+import { OutputFormat } from '../types.js';
 import { baseService, LocalImageService } from './service.js';
 
 let sharp: typeof import('sharp');
@@ -10,23 +10,7 @@ try {
 
 const sharpService: LocalImageService = {
 	getURL: baseService.getURL,
-	parseParams(params) {
-		if (!params.has('href')) {
-			return undefined;
-		}
-
-		let transform: Partial<ImageTransform> = { src: params.get('href')! };
-
-		if (params.has('w')) {
-			transform.width = parseInt(params.get('w')!);
-		}
-
-		if (params.has('h')) {
-			transform.height = parseInt(params.get('h')!);
-		}
-
-		return transform;
-	},
+	parseParams: baseService.parseParams,
 	async transform(inputBuffer, transform) {
 		// If the user didn't specify a format, we'll default to `webp`. It offers the best ratio of compatibility / quality
 		// In the future, hopefully we can replace this with `avif`, alas, Edge. See https://caniuse.com/avif
